src/requirements.txt
chromadb
langchain
langchain_ollama
langchain_core
fastapi
uvicorn
openai
aiofiles
python-multipart
pymupdf
email-validator
